Extract utils for std::vector<double>
